replace content length with long instead of int

+++ src/main/java/com/perimeterx/http/PXRequestBody.java
@@ -9,5 +9,5 @@ import java.io.InputStream;
 @AllArgsConstructor
 public class PXRequestBody {
     private InputStream inputStream;
-    private int length;
+    private long length;
 }
